refactor: extract shared CORS config to _shared/cors.ts
