+++ supabase/functions/auth-github-token/index.ts
@@ -27,71 +27,13 @@ import {
   create,
   getNumericDate,
 } from 'https://deno.land/x/djwt@v3.0.2/mod.ts';
+import { handleCors } from '../_shared/cors.ts';
 
 // =============================================================================
 // Constants
 // =============================================================================
 
-const AUTH_GITHUB_VERSION = '1.0.1';
-
-/**
- * Allowed CORS origins for security.
- * VS Code extensions use opaque origins so we must allow those schemes.
- * Codespaces uses *.github.dev domains.
- */
-const ALLOWED_ORIGINS = [
-  // VS Code and forks (opaque origins, sent as null in most browsers)
-  'vscode://',
-  'vscode-insiders://',
-  'cursor://',
-  'windsurf://',
-  'antigravity://',
-  // Codespaces and github.dev
-  /^https:\/\/[a-z0-9-]+\.github\.dev$/,
-  /^https:\/\/[a-z0-9-]+\.app\.github\.dev$/,
-  // TeXRA domains
-  /^https:\/\/([a-z0-9-]+\.)?texra\.ai$/,
-  // localhost for development
-  /^http:\/\/localhost(:\d+)?$/,
-];
-
-/**
- * Check if origin is allowed for CORS.
- * Returns the origin if allowed, null otherwise.
- */
-function getAllowedOrigin(origin: string | null): string | null {
-  if (!origin) {
-    // Null origin (from opaque origins like vscode://) - allow for VS Code extensions
-    return '*';
-  }
-
-  for (const allowed of ALLOWED_ORIGINS) {
-    if (typeof allowed === 'string') {
-      if (origin.startsWith(allowed)) {
-        return origin;
-      }
-    } else if (allowed.test(origin)) {
-      return origin;
-    }
-  }
-
-  return null;
-}
-
-/**
- * Get CORS headers for a request.
- */
-function getCorsHeaders(req: Request): Record<string, string> {
-  const origin = req.headers.get('Origin');
-  const allowedOrigin = getAllowedOrigin(origin);
-
-  return {
-    'Access-Control-Allow-Origin': allowedOrigin || '',
-    'Access-Control-Allow-Headers': 'authorization, x-client-info, apikey, content-type',
-    'Access-Control-Allow-Methods': 'POST, OPTIONS',
-    'Vary': 'Origin',
-  };
-}
+const AUTH_GITHUB_VERSION = '1.0.2';
 
 // Session duration: 1 hour access token, 7 day refresh token
 const ACCESS_TOKEN_EXPIRY_SECONDS = 3600;
@@ -458,18 +400,9 @@ if (!supabaseUrl || !supabaseServiceKey || !jwtSecret) {
 // =============================================================================
 
 Deno.serve(async (req: Request) => {
-  // Get dynamic CORS headers based on request origin
-  const corsHeaders = getCorsHeaders(req);
-
-  // Reject requests from disallowed origins
-  if (!corsHeaders['Access-Control-Allow-Origin']) {
-    return new Response('Forbidden', { status: 403 });
-  }
-
-  // Handle CORS preflight
-  if (req.method === 'OPTIONS') {
-    return new Response('ok', { headers: corsHeaders });
-  }
+  // Handle CORS (preflight, origin validation)
+  const { corsHeaders, response: corsResponse } = handleCors(req);
+  if (corsResponse) return corsResponse;
 
   // Only accept POST requests
   if (req.method !== 'POST') {
